+++ esp/zicBlock/main/engineView.h
@@ -49,7 +49,7 @@ UIManagerInterface& ui;
         // renderStringValue(valuePos[8], "Reverb", std::to_string((int)(audio.metalic.reverb * 100)) + "%");
     }
 
-    void onEncoder(int id, int8_t direction, uint64_t tick) override
+    void onEncoder(int8_t id, int8_t direction, uint64_t tick) override
     {
         if (id == 1) {
             ui.selectEngine(ui.getSelectedEngine() + (direction > 0 ? 1 : -1));
