+++ esp/zicBlock/main/uiManager.h
@@ -111,7 +111,7 @@ public:
     }
 
     uint64_t lastEncoderTick[encoderCount] = { 0 };
-    void onEncoder(int id, int8_t direction, uint64_t tick)
+    void onEncoder(int8_t id, int8_t direction, uint64_t tick)
     {
         int scaledDirection = encGetScaledDirection(direction, tick, lastEncoderTick[id]);
         lastEncoderTick[id] = tick;
