+++ plugins/components/Pixel/FmAlgoComponent.h
@@ -175,7 +175,7 @@ public:
         }
     }
 
-    void onEncoder(int id, int8_t direction)
+    void onEncoder(int8_t id, int8_t direction)
     {
         if (value && id == encoderId) {
             value->increment(direction);
