+++ plugins/components/Pixel/GitHubComponent.h
@@ -616,7 +616,7 @@ public:
         }
     }
 
-    void onEncoder(int id, int8_t direction)
+    void onEncoder(int8_t id, int8_t direction)
     {
         if (id == encoderId && reposLoaded && !repos.empty()) {
             currentRepoIndex += direction;
