+++ plugins/components/Pixel/NoteGridComponent.h
@@ -128,7 +128,7 @@ public:
         updateNotes();
     }
 
-    void onEncoder(int id, int8_t direction) override
+    void onEncoder(int8_t id, int8_t direction) override
     {
         direction = direction > 0 ? 1 : -1;
         if (id == encScale) {
