+++ plugins/components/Pixel/WorkspaceKnobComponent.h
@@ -125,7 +125,7 @@ public:
         }
     }
 
-    void onEncoder(int id, int8_t direction) override
+    void onEncoder(int8_t id, int8_t direction) override
     {
         if (id == encoderId) {
             workspace = CLAMP(workspace + direction, 1, 99);
