+++ plugins/components/__deprecated_SDL/EncoderBarComponent.h
@@ -236,7 +236,7 @@ public:
         return false;
     }
 
-    void onEncoder(int id, int8_t direction)
+    void onEncoder(int8_t id, int8_t direction)
     {
         if (id == encoderId) {
             value->increment(direction);
